Add files via upload
We are first finding out the next greater number of each number in our array. Then for a window, we check if the element has the next greatest element within the window or not. If yes, then we move on to check if that element's next greatest element is within the window and continue to check for the same till the next greatest element for the numbers within the window, lies outside the window. When this is established, it means that all the numbers between the next greatest element and its corresponding number in the window, are all smaller to the number. Thus this number is the greatest number in the window. 

We do this procedure to complete the procedure in the O(n) complexity.